Move manifest to new namespace
This patch moves the `Manifest` struct and additional items to a new
sub namespace `v1`. The structs get reexported as before to not break
the current API.
